renamed db_query
renaming db_query to avoid errors
