+++ src/pe_source/dnstwistscript.py
@@ -6,7 +6,7 @@ import logging
 import traceback
 
 # Third-Party Libraries
-from data.pe_db.db_query import (
+from data.pe_db.db_query_source import (
     addSubdomain,
     connect,
     getDataSource,
